[Backport] Security bug 905509 (9/13)
Backport of original patch by Antoine Labour <piman@chromium.org>:
Audit size_t: odds and ends in gpu/command_buffer/client

Mostly just replacing size_t (and unsigned int) to uint32_t as
appropriate.

Bug: 905509
Reviewed-on: https://chromium-review.googlesource.com/c/1401850

+++ chromium/gpu/command_buffer/client/transfer_buffer_unittest.cc
@@ -33,9 +33,9 @@ class TransferBufferTest : public testing::Test {
   static const int32_t kNumCommandEntries = 400;
   static const int32_t kCommandBufferSizeBytes =
       kNumCommandEntries * sizeof(CommandBufferEntry);
-  static const unsigned int kStartingOffset = 64;
-  static const unsigned int kAlignment = 4;
-  static const size_t kTransferBufferSize = 256;
+  static const uint32_t kStartingOffset = 64;
+  static const uint32_t kAlignment = 4;
+  static const uint32_t kTransferBufferSize = 256;
 
   TransferBufferTest()
       : transfer_buffer_id_(0) {
@@ -91,9 +91,9 @@ void TransferBufferTest::TearDown() {
 #ifndef _MSC_VER
 const int32_t TransferBufferTest::kNumCommandEntries;
 const int32_t TransferBufferTest::kCommandBufferSizeBytes;
-const unsigned int TransferBufferTest::kStartingOffset;
-const unsigned int TransferBufferTest::kAlignment;
-const size_t TransferBufferTest::kTransferBufferSize;
+const uint32_t TransferBufferTest::kStartingOffset;
+const uint32_t TransferBufferTest::kAlignment;
+const uint32_t TransferBufferTest::kTransferBufferSize;
 #endif
 
 TEST_F(TransferBufferTest, Basic) {
@@ -158,7 +158,7 @@ TEST_F(TransferBufferTest, Free) {
   EXPECT_EQ(base::UnguessableToken(), transfer_buffer_->shared_memory_guid());
 
   // See that it gets reallocated.
-  unsigned int size = 0;
+  uint32_t size = 0;
   void* data = transfer_buffer_->AllocUpTo(1, &size);
   EXPECT_TRUE(data != NULL);
   EXPECT_TRUE(transfer_buffer_->HaveBuffer());
@@ -205,7 +205,7 @@ TEST_F(TransferBufferTest, TooLargeAllocation) {
   void* ptr = transfer_buffer_->Alloc(kTransferBufferSize + 1);
   EXPECT_TRUE(ptr == NULL);
   // Check we if we try to allocate larger than max we get max.
-  unsigned int size_allocated = 0;
+  uint32_t size_allocated = 0;
   ptr = transfer_buffer_->AllocUpTo(
       kTransferBufferSize + 1, &size_allocated);
   ASSERT_TRUE(ptr != NULL);
@@ -243,11 +243,11 @@ class TransferBufferExpandContractTest : public testing::Test {
   static const int32_t kNumCommandEntries = 400;
   static const int32_t kCommandBufferSizeBytes =
       kNumCommandEntries * sizeof(CommandBufferEntry);
-  static const unsigned int kStartingOffset = 64;
-  static const unsigned int kAlignment = 4;
-  static const size_t kStartTransferBufferSize = 256;
-  static const size_t kMaxTransferBufferSize = 1024;
-  static const size_t kMinTransferBufferSize = 128;
+  static const uint32_t kStartingOffset = 64;
+  static const uint32_t kAlignment = 4;
+  static const uint32_t kStartTransferBufferSize = 256;
+  static const uint32_t kMaxTransferBufferSize = 1024;
+  static const uint32_t kMinTransferBufferSize = 128;
 
   TransferBufferExpandContractTest()
       : transfer_buffer_id_(0) {
@@ -320,11 +320,11 @@ void TransferBufferExpandContractTest::TearDown() {
 #ifndef _MSC_VER
 const int32_t TransferBufferExpandContractTest::kNumCommandEntries;
 const int32_t TransferBufferExpandContractTest::kCommandBufferSizeBytes;
-const unsigned int TransferBufferExpandContractTest::kStartingOffset;
-const unsigned int TransferBufferExpandContractTest::kAlignment;
-const size_t TransferBufferExpandContractTest::kStartTransferBufferSize;
-const size_t TransferBufferExpandContractTest::kMaxTransferBufferSize;
-const size_t TransferBufferExpandContractTest::kMinTransferBufferSize;
+const uint32_t TransferBufferExpandContractTest::kStartingOffset;
+const uint32_t TransferBufferExpandContractTest::kAlignment;
+const uint32_t TransferBufferExpandContractTest::kStartTransferBufferSize;
+const uint32_t TransferBufferExpandContractTest::kMaxTransferBufferSize;
+const uint32_t TransferBufferExpandContractTest::kMinTransferBufferSize;
 #endif
 
 TEST_F(TransferBufferExpandContractTest, ExpandWithSmallAllocations) {
@@ -351,7 +351,7 @@ TEST_F(TransferBufferExpandContractTest, ExpandWithSmallAllocations) {
       transfer_buffer_->GetCurrentMaxAllocationWithoutRealloc());
 
   // Fill the free space.
-  unsigned int size_allocated = 0;
+  uint32_t size_allocated = 0;
   void* ptr = transfer_buffer_->AllocUpTo(transfer_buffer_->GetFreeSize(),
                                           &size_allocated);
   transfer_buffer_->FreePendingToken(ptr, token);
@@ -417,7 +417,7 @@ TEST_F(TransferBufferExpandContractTest, ExpandWithLargeAllocations) {
 
   // Expand again.
   ExpectCreateTransferBuffer(kStartTransferBufferSize * 4);
-  unsigned int size_requested = transfer_buffer_->GetFreeSize() + 1;
+  uint32_t size_requested = transfer_buffer_->GetFreeSize() + 1;
   ptr = transfer_buffer_->AllocUpTo(size_requested, &size_allocated);
   ASSERT_TRUE(ptr != NULL);
   EXPECT_EQ(size_requested, size_allocated);
@@ -461,7 +461,7 @@ TEST_F(TransferBufferExpandContractTest, ShrinkRingBuffer) {
   transfer_buffer_->FreePendingToken(ptr, token);
 
   // We shouldn't shrink before we reach the allocation threshold.
-  for (size_t allocated = kMaxTransferBufferSize - kStartingOffset;
+  for (uint32_t allocated = kMaxTransferBufferSize - kStartingOffset;
        allocated < (kStartTransferBufferSize + kStartingOffset) *
                        (TransferBuffer::kShrinkThreshold);) {
     ptr = transfer_buffer_->Alloc(kStartTransferBufferSize);
@@ -506,9 +506,9 @@ TEST_F(TransferBufferExpandContractTest, Contract) {
           &MockClientCommandBufferCanFail::RealCreateTransferBuffer))
       .RetiresOnSaturation();
 
-  const size_t kSize1 = 256 - kStartingOffset;
-  const size_t kSize2 = 128 - kStartingOffset;
-  unsigned int size_allocated = 0;
+  const uint32_t kSize1 = 256 - kStartingOffset;
+  const uint32_t kSize2 = 128 - kStartingOffset;
+  uint32_t size_allocated = 0;
   void* ptr = transfer_buffer_->AllocUpTo(kSize1, &size_allocated);
   ASSERT_TRUE(ptr != NULL);
   EXPECT_EQ(kSize2, size_allocated);
@@ -563,8 +563,8 @@ TEST_F(TransferBufferExpandContractTest, OutOfMemory) {
            DoAll(SetArgPointee<1>(-1), Return(scoped_refptr<gpu::Buffer>())))
       .RetiresOnSaturation();
 
-  const size_t kSize1 = 512 - kStartingOffset;
-  unsigned int size_allocated = 0;
+  const uint32_t kSize1 = 512 - kStartingOffset;
+  uint32_t size_allocated = 0;
   void* ptr = transfer_buffer_->AllocUpTo(kSize1, &size_allocated);
   ASSERT_TRUE(ptr == NULL);
   EXPECT_FALSE(transfer_buffer_->HaveBuffer());
@@ -599,9 +599,9 @@ TEST_F(TransferBufferExpandContractTest, ReallocsToDefault) {
 }
 
 TEST_F(TransferBufferExpandContractTest, Shrink) {
-  unsigned int alloc_size = transfer_buffer_->GetFreeSize();
+  uint32_t alloc_size = transfer_buffer_->GetFreeSize();
   EXPECT_EQ(kStartTransferBufferSize - kStartingOffset, alloc_size);
-  unsigned int size_allocated = 0;
+  uint32_t size_allocated = 0;
   void* ptr = transfer_buffer_->AllocUpTo(alloc_size, &size_allocated);
 
   ASSERT_NE(ptr, nullptr);
@@ -610,13 +610,13 @@ TEST_F(TransferBufferExpandContractTest, Shrink) {
   EXPECT_EQ(0u, transfer_buffer_->GetFreeSize());
 
   // Shrink once.
-  const unsigned int shrink_size1 = 64;
+  const uint32_t shrink_size1 = 64;
   EXPECT_LT(shrink_size1, alloc_size);
   transfer_buffer_->ShrinkLastBlock(shrink_size1 - kAlignment + 1);
   EXPECT_EQ(alloc_size - shrink_size1, transfer_buffer_->GetFreeSize());
 
   // Shrink again.
-  const unsigned int shrink_size2 = 32;
+  const uint32_t shrink_size2 = 32;
   EXPECT_LT(shrink_size2, shrink_size1);
   transfer_buffer_->ShrinkLastBlock(shrink_size2);
   EXPECT_EQ(alloc_size - shrink_size2, transfer_buffer_->GetFreeSize());
